requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.14.0
python-dotenv==1.1.1
requests==2.32.5
firebase-admin==6.5.0
pydantic==2.9.0